Permissions

See merge request adrien.paysant/booker!25

+++ Booker/src/main/java/ch/hearc/qdljee/controllers/BooksController.java
@@ -5,6 +5,7 @@ import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.sql.Date;
+import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -34,6 +35,7 @@ import ch.hearc.qdljee.dto.CommentDto;
 import ch.hearc.qdljee.dto.RatingsDto;
 import ch.hearc.qdljee.dto.SearchDto;
 import ch.hearc.qdljee.model.Books;
+import ch.hearc.qdljee.model.Role;
 import ch.hearc.qdljee.model.User;
 import ch.hearc.qdljee.service.BookService;
 import ch.hearc.qdljee.service.CommentService;
@@ -156,8 +158,11 @@ public class BooksController {
 
 	@PostMapping("/{id}/delete")
 	public String deleteBook(Model model, @PathVariable("id") int id) {
+		if (!validatePermForBook(id)) {
+			return "redirect:/Books/?errorPermissions";
+		}
 		String path = environment.getProperty("images.path");
-		bookService.delete(id,path);
+		bookService.delete(id, path);
 		return "redirect:/Books";
 	}
 
@@ -170,6 +175,10 @@ public class BooksController {
 
 	@PostMapping("/{id}/update")
 	public String updateBook(Model model, @PathVariable("id") int id, @ModelAttribute("Bookdto") BookDto bookDto) {
+		if (!validatePermForBook(id)) {
+			return "redirect:/Books/?errorPermissions";
+		}
+
 		String imageURL = null;
 		if (bookDto.getImage() != null && bookDto.getImage().getOriginalFilename().contains("octet-stream")) {
 			String path = environment.getProperty("images.path");
@@ -213,6 +222,9 @@ public class BooksController {
 	@PostMapping("/{id}/comment/{comId}/update")
 	public String updateComment(Model model, @PathVariable("id") int id, @PathVariable("comId") int comId,
 			@ModelAttribute("CommentDto") CommentDto comDto) throws Exception {
+		if (!validatePermForComment(comId)) {
+			return "redirect:/Books/" + id + "?errorPermissions";
+		}
 		comDto.setPublicationDate(new Date(System.currentTimeMillis()));
 		comDto.setBookId(id);
 		commentService.update(comDto, comId);
@@ -221,7 +233,39 @@ public class BooksController {
 
 	@PostMapping("/{id}/comment/{comId}/delete")
 	public String deleteComment(Model model, @PathVariable("id") int id, @PathVariable("comId") int comId) {
+		if (!validatePermForComment(comId)) {
+			return "redirect:/Books/" + id + "?errorPermissions";
+		}
 		commentService.delete(comId);
 		return "redirect:/Books/" + id;
+	}
+
+	private boolean validatePermForComment(int comId) {
+		Collection<Role> roles = Tools.getCurrentUser().getRoles();
+		boolean status = false;
+		for (Role role : roles) {
+			if (commentService.getCommentsById(comId).getAuthor().equals(Tools.getCurrentUser())) {
+				status = true;
+			}
+			if (role.getName().equals("ROLE_ADMIN")) {
+				status = false;
+			}
+		}
+		return !status;
+	}
+
+	private boolean validatePermForBook(int id) {
+		Collection<Role> roles = Tools.getCurrentUser().getRoles();
+		boolean status = false;
+		for (Role role : roles) {
+			if (!role.getName().equals("ROLE_AUTHOR")
+					&& !bookService.getBooksById(id).getAuthor().equals(Tools.getCurrentUser())) {
+				status = true;
+			}
+			if (role.getName().equals("ROLE_ADMIN")) {
+				status = false;
+			}
+		}
+		return !status;
 	}
 }
